Updating to current Stu-FE schema
Committed via web, not via GithubGUI (it's giving me #$%* on the backend branch).
This does overwrite some/many of Philip's changes, but it's synced with what I'm working on (and what I need to be tested via grunt test)
I'm working on consolidating the two schema (schemas? schemae? schemi? yay Latin!...or is it Greek?)